Own CMakeList.txt so it can go in 3rdparty. Fixed Guide types.

+++ telescope_sitech.h
@@ -18,12 +18,12 @@
 
 #ifndef SCOPESITECH_H
 #define SCOPESITECH_H
-#include "indibase/indiguiderinterface.h"
-#include "indibase/inditelescope.h"
-#include "indicontroller.h"
-#include "indidevapi.h"
-#include "indicom.h"
-#include "indibase/baseclient.h"
+#include <indiguiderinterface.h>
+#include <inditelescope.h>
+#include <indicontroller.h>
+#include <indidevapi.h>
+#include <indicom.h>
+#include <baseclient.h>
 class ScopeSiTech : public INDI::Telescope, public INDI::GuiderInterface
 {
 public:
@@ -48,10 +48,10 @@ public:
     virtual bool MoveWE(INDI_DIR_WE dir, TelescopeMotionCommand command);
     virtual bool Abort();
 
-    virtual IPState GuideNorth(float ms);
-    virtual IPState GuideSouth(float ms);
-    virtual IPState GuideEast(float ms);
-    virtual IPState GuideWest(float ms);
+    virtual IPState GuideNorth(uint32_t ms);
+    virtual IPState GuideSouth(uint32_t ms);
+    virtual IPState GuideEast(uint32_t ms);
+    virtual IPState GuideWest(uint32_t ms);
 
     bool Goto(double,double);
     bool Park();
